Activate Profile (#73)
* Redirect to my user profile instead.

* Add a profile required decorator to trigger activation.

* Setup the optin marketing flag.

* Some simple messaging.

* Add the profile required middleware where required.

* New decorator and some code cleanup.

* Core helpers.

* Change decorator usage.

* Style the profile page.

* Style the profile page.

+++ play/apps/core/views/profile.py
@@ -2,33 +2,37 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth import logout
+from apps.core.models import Profile
 from apps.core.forms import ProfileForm
-from apps.core.middleware import with_profile
+from apps.core.middleware import profile_required
 
 
 @login_required
-@with_profile
 def edit(request):
-    form = ProfileForm(instance=request.user.profile)
-    return render(request, "profile/edit.html", {"form": form})
+    profile = Profile.objects.get_or_init(user=request.user)
+    form = ProfileForm(instance=profile)
+    return render(request, "profile/edit.html", {"form": form, "profile": profile})
 
 
 @login_required
-@with_profile
 def update(request):
-    form = ProfileForm(request.POST, instance=request.user.profile)
+    profile = Profile.objects.get_or_init(user=request.user)
+    form = ProfileForm(request.POST, instance=profile)
     if form.is_valid():
         form.save()
         messages.add_message(request, messages.INFO, "Updated profile")
         return redirect("profile")
-    return render(request, "profile/edit.html", {"form": form}, status=400)
+    return render(
+        request, "profile/edit.html", {"form": form, "profile": profile}, status=400
+    )
 
 
 @login_required
-@with_profile
+@profile_required
 def delete(request):
     user = request.user
     logout(request)
     user.delete()
-    user.profile.delete()
+    if hasattr(user, "profile"):
+        user.profile.delete()
     return redirect("/")
